[SPARK-17661][SQL] Consolidate various listLeafFiles implementations
